Added output of pressure values

+++ app/src/main/java/open/sqxx/vostok27/presentation/main/sensors/SensorsView.kt
@@ -6,5 +6,5 @@ interface SensorsView : MvpView {
 	fun showCO2(value: Int)
 	fun showTemp(value: Int)
 	fun showHumidity(value: Int)
-	fun showPressure(value: Int)
+	fun showPressure(value: Float)
 }
